clean and expand
Applies these learnings:

If all we do is store an index matrix from `max` into Policy, we don't need to cast it to the index type.  However, if we need to do any maths, we do.

If we do any maths with division within a rem/ceil stanza, we need to take care that we are promoting numerator and denominator to double.  This is because in the integer world ceil(1/2) is zero, whereas ceil(1.0/2.0) is 1.  You'll know you got it wrong when you find lots of zeros in the Policy matrix.

Expanded to cover the Huggett&Ventura 2000 replication.

+++ StationaryDist/FHorz/RiskyAsset/StationaryDist_FHorz_RiskyAsset.m
@@ -6,7 +6,7 @@ if isUnderlyingType(jequaloneDist,'single')
     precision_cast=@(x) single(x);
 else
     precision='double';
-    precision_index='int32';
+    precision_index='double';
     precision_cast=@(x) double(x);
 end
 
@@ -165,7 +165,7 @@ elseif simoptions.gridinterplayer==1
     % Policy_aprime(:,:,:,1:2*N_u,:) lower grid point for a1 is unchanged
     Policy_aprime(:,:,2*N_u+1:end,:)=Policy_aprime(:,:,2*N_u+1:end,:)+1; % add one to a1, to get upper grid point
 
-    aprimeProbs_upper=reshape(shiftdim((Policy(end-1,:,:,:)-1)/(simoptions.ngridinterp+1),1),[N_a,N_ze,1,N_j]); % probability of upper grid point (from L2 index; end-1 because end is now L2flag)
+    aprimeProbs_upper=reshape(shiftdim(double(Policy(end-1,:,:,:)-1)/(simoptions.ngridinterp+1),1),[N_a,N_ze,1,N_j]); % probability of upper grid point (from L2 index; end-1 because end is now L2flag)
     PolicyProbs(:,:,1:2*N_u,:)=PolicyProbs(:,:,1:2*N_u,:).*(1-aprimeProbs_upper); % lower a1
     PolicyProbs(:,:,2*N_u+1:end,:)=PolicyProbs(:,:,2*N_u+1:end,:).*aprimeProbs_upper; % upper a1
